refactor(api): api client generico
Sviluppato un modulo denominato client, in grado di importare dinamicamente librerie client i cui metodi possono essere invocati a runtime

#9
