+++ packages/.env
@@ -1,8 +1,4 @@
-PUBLIC_URL_API_FANTAMASTER=/api/my/api/fantamaster
-PUBLIC_URL_PLAYERS=/api/my/api/players
-PUBLIC_URL_STATS=/api/my/fantamaster/season_stats
-PUBLIC_URL_UNAVAILABLE_PLAYERS=/api/my/fantamaster/unavailable
-PUBLIC_URL_LINEUP=/api/my/fantamaster/lineup
+PUBLIC_URL_API_BASEURL=/api/my/api/client
 
 PUBLIC_URL_AI_PRESENTATION=/api/my/chatbot/presentation
 PUBLIC_URL_AI_SINGLE_PLAYER=/api/my/chatbot/single_player
